Move some IQ declaration in their own files
